Move genericManifest-related code to image/manifest.go
image/manifest.go now contains genericManifest without caring about
genericImage, and image/image.go contains genericImage only.
